Moved the base VICE admin page into pages from the components. Prevents future entanglements between components
